requirements.txt
python-docx 